Extract Model types out of the Api
