+++ Services/CO.CDP.DataSharing.WebApi/Api/DataSharing.cs
@@ -1,115 +1,10 @@
-using System.ComponentModel.DataAnnotations;
-using Microsoft.OpenApi.Models;
+using System.Reflection;
+using CO.CDP.DataSharing.WebApi.Model;
 using DotSwashbuckle.AspNetCore.SwaggerGen;
+using Microsoft.OpenApi.Models;
 
 namespace CO.CDP.DataSharing.WebApi.Api
 {
-    internal record SupplierInformation
-    {
-        [Required] public required Guid Id { get; init; }
-        [Required(AllowEmptyStrings = true)] public required string Name { get; init; }
-        [Required] public required List<OrganisationReference> AdditionalParties { get; init; }
-        [Required] public required Identifier Identifier { get; init; }
-        [Required] public required List<Identifier> AdditionalIdentifiers { get; init; }
-        [Required] public required Address Address { get; init; }
-        [Required] public required List<ContactPoint> ContactPoints { get; init; }
-        [Required] public required List<PartyRole> Roles { get; init; }
-        [Required(AllowEmptyStrings = true)] public required string Details { get; init; }
-        [Required] public required SupplierInformationData SupplierInformationData { get; init; }
-    }
-
-    internal record OrganisationReference
-    {
-        [Required] public required Guid Id { get; init; }
-        [Required(AllowEmptyStrings = true)] public required string Name { get; init; }
-    }
-
-    internal record Identifier
-    {
-        [Required(AllowEmptyStrings = true)] public required string Scheme { get; init; }
-        [Required] public required Guid Id { get; init; }
-        public string? LegalName { get; init; }
-        public string? Uri { get; init; }
-    }
-
-    internal record Address
-    {
-        [Required(AllowEmptyStrings = true)] public required string StreetAddress { get; init; }
-        [Required(AllowEmptyStrings = true)] public required string Locality { get; init; }
-        public string? Region { get; init; }
-        public string? PostalCode { get; init; }
-    }
-
-    internal record ContactPoint
-    {
-        public string? Email { get; init; }
-        public string? Telephone { get; init; }
-        public string? FaxNumber { get; init; }
-    }
-
-    internal enum PartyRole
-    {
-        Buyer,
-        ProcuringEntity,
-        Supplier,
-        Tenderer,
-        Funder,
-        Enquirer,
-        Payer,
-        Payee,
-        ReviewBody,
-        InterestedParty,
-        Consortium
-    }
-
-    internal record SupplierInformationData
-    {
-        [Required] public required Guid FormId { get; init; }
-        public List<SupplierFormAnswer> Answers { get; init; } = new List<SupplierFormAnswer>();
-        public List<FormQuestion> Questions { get; init; } = new List<FormQuestion>();
-
-    }
-
-    internal record SupplierFormAnswer
-    {
-        public string? QuestionName { get; init; }
-        public bool? BoolValue { get; init; }
-        public double? NumericValue { get; init; }
-        public DateTime? StartValue { get; init; }
-        public DateTime? EndValue { get; init; }
-        public string? TextValue { get; init; }
-        public int? OptionValue { get; init; }
-    }
-
-    internal record FormQuestion
-    {
-        public QuestionTypes? Type { get; init; }
-        public string? Name { get; init; }
-        public string? Text { get; init; }
-        public bool IsRequired { get; init; }
-        public string? SectionName { get; init; }
-        public List<QuestionOption> Options { get; init; } = new List<QuestionOption>();
-    }
-
-    internal enum QuestionTypes
-    {
-        Type1 = 1,
-        Type2 = 2,
-        Type3 = 3,
-        Type4 = 4,
-        Type5 = 5,
-        Type6 = 6,
-        Type7 = 7,
-        Type8 = 8,
-        Type9 = 9
-    }
-
-    internal record QuestionOption
-    {
-        [Required] public required Guid Id { get; init; }
-        public string? Value { get; init; }
-    }
-
     public static class EndpointExtensions
     {
         public static void UseDataSharingEndpoints(this WebApplication app)
