+++ Services/CO.CDP.DataSharing.WebApi/Model/FormQuestion.cs
@@ -1,0 +1,11 @@
+namespace CO.CDP.DataSharing.WebApi.Model;
+
+internal record FormQuestion
+{
+    public QuestionTypes? Type { get; init; }
+    public string? Name { get; init; }
+    public string? Text { get; init; }
+    public bool IsRequired { get; init; }
+    public string? SectionName { get; init; }
+    public List<QuestionOption> Options { get; init; } = new List<QuestionOption>();
+}
