+++ Services/CO.CDP.DataSharing.WebApi/Model/OrganisationReference.cs
@@ -1,0 +1,9 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace CO.CDP.DataSharing.WebApi.Model;
+
+internal record OrganisationReference
+{
+    [Required] public required Guid Id { get; init; }
+    [Required(AllowEmptyStrings = true)] public required string Name { get; init; }
+}
