+++ Services/CO.CDP.DataSharing.WebApi/Model/PartyRole.cs
@@ -1,0 +1,16 @@
+namespace CO.CDP.DataSharing.WebApi.Model;
+
+internal enum PartyRole
+{
+    Buyer,
+    ProcuringEntity,
+    Supplier,
+    Tenderer,
+    Funder,
+    Enquirer,
+    Payer,
+    Payee,
+    ReviewBody,
+    InterestedParty,
+    Consortium
+}
