+++ Services/CO.CDP.DataSharing.WebApi/Model/QuestionTypes.cs
@@ -1,0 +1,14 @@
+namespace CO.CDP.DataSharing.WebApi.Model;
+
+internal enum QuestionTypes
+{
+    Type1 = 1,
+    Type2 = 2,
+    Type3 = 3,
+    Type4 = 4,
+    Type5 = 5,
+    Type6 = 6,
+    Type7 = 7,
+    Type8 = 8,
+    Type9 = 9
+}
