+++ Services/CO.CDP.DataSharing.WebApi/Model/SupplierFormAnswer.cs
@@ -1,0 +1,12 @@
+namespace CO.CDP.DataSharing.WebApi.Model;
+
+internal record SupplierFormAnswer
+{
+    public string? QuestionName { get; init; }
+    public bool? BoolValue { get; init; }
+    public double? NumericValue { get; init; }
+    public DateTime? StartValue { get; init; }
+    public DateTime? EndValue { get; init; }
+    public string? TextValue { get; init; }
+    public int? OptionValue { get; init; }
+}
